// src/store/notificationStore/action.enum.ts
enum actionTypes {

  GET_NOTIFICATIONS_SUCCESS = "GET_NOTIFICATIONS_SUCCESS",
  GET_NOTIFICATIONS_PENDING = "GET_NOTIFICATIONS_PENDING",
  GET_NOTIFICATIONS_FAILED = "GET_NOTIFICATIONS_FAILED",

  GET_HEADER_NOTIFICATIONS_SUCCESS = "GET_HEADER_NOTIFICATIONS_SUCCESS",
  GET_HEADER_NOTIFICATIONS_PENDING = "GET_HEADER_NOTIFICATIONS_PENDING",
  GET_HEADER_NOTIFICATIONS_FAILED = "GET_HEADER_NOTIFICATIONS_FAILED",

  DELETE_NOTIFICATIONS_SUCCESS = "DELETE_NOTIFICATIONS_SUCCESS",
  DELETE_NOTIFICATIONS_PENDING = "DELETE_NOTIFICATIONS_PENDING",
  DELETE_NOTIFICATIONS_FAILED = "DELETE_NOTIFICATIONS_FAILED",

  READ_NOTIFICATIONS_SUCCESS = "READ_NOTIFICATIONS_SUCCESS",
  READ_NOTIFICATIONS_PENDING = "READ_NOTIFICATIONS_PENDING",
  READ_NOTIFICATIONS_FAILED = "READ_NOTIFICATIONS_FAILED",

  ADD_NEW_NOTIFICATION = "ADD_NEW_NOTIFICATION"
}

export default actionTypes;